+++ test/tools/add_typescript_version_tag_test.dart
@@ -245,5 +245,10 @@ void main() {
         isA<AddTypeScriptVersionTag>(),
       );
     });
+
+    // .........................................................................
+    test('.example builds a usable instance', () {
+      expect(AddTypeScriptVersionTag.example(), isA<AddTypeScriptVersionTag>());
+    });
   });
 }