+++ test/tools/checks_test.dart
@@ -16,12 +16,13 @@ void main() {
   group('Checks', () {
     group('all', () {
       test('should provide a list of all checks', () {
-        expect(checks.all, hasLength(10));
+        expect(checks.all, hasLength(11));
         expect(checks.all, [
           checks.pubGetOffline,
           checks.analyze,
           checks.format,
           checks.tests,
+          checks.packageJsonScripts,
           checks.pana,
           checks.isPushed,
           checks.isCommitted,